Implemented class inheritance in perf tests to reduce overlapping code
